+++ packages/validate/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  validateHtmlInBrowser,
+  type BrowserValidateOptions,
+  type ConsoleEntry,
+  type ContrastEntry,
+} from "./browserValidate.js";
